move GuidManager to util directory
